requirements.txt
streamlit
google-cloud-bigquery==3.3.6